This resolves #42.
unbundled-bezier, used in 2 classes (one for each tier of bowing) is used when a gap junction is present to prevent overlap.

Note that each tier is meant to contain 1 connection type where possible.  Default in the full case is synapses on the inner tier and funcons in the outer.

Ostensibly, this should have been able to work with a single class with 4 control-point-distances in a string, but every time I tried that, there was only 1 tier of bowing and overlap would occur.

Files checked in: src/client/js/view/network-style.js src/client/js/view/network.js